mito-ai: small refactor
